On to the next version (#47)
A complete rewrite.
Uses `acorn` to parse function in order to get more detailed output.

BREAKING CHANGE: the representation has changed.

Now it cannot be parsed back to object using simple javascript parser because I shorten some of the keywords (e.g. `function` to `fn`, `Symbol` to `Sym`).

This change is justified as `tersify` is primary a utility to create a representation of value that is readable by a human.

If you are using it with `maxLength: Infinity` to stringify an object, use `flatten` instead.
It does a much better job than this naive code.